rider40: add correction for SW bug in my Rider40
This manifests as a FAST point (to close together, factor of 2) followed
shortly after by a SLOW point (to far apart, factor of 2) and then a missing
GPS point.

I think the SW is logging a late GPS point against the wrong timestamp then
continuing on that time line for a few points before correcting itself (by
dropping a point).